rand and randomUniform
