+++ tfjs-core/src/ops/array_ops_test.ts
@@ -17,8 +17,7 @@
 
 import * as tf from '../index';
 import {ALL_ENVS, BROWSER_ENVS, describeWithFlags, NODE_ENVS} from '../jasmine_util';
-import {expectArraysClose, expectArraysEqual, expectPromiseToFail, expectValuesInRange} from '../test_util';
-import * as util from '../util';
+import {expectArraysClose, expectArraysEqual, expectPromiseToFail} from '../test_util';
 
 describeWithFlags('zeros', ALL_ENVS, () => {
   it('1D default dtype', async () => {
@@ -764,116 +763,6 @@ describeWithFlags('onesLike', ALL_ENVS, () => {
   });
 });
 
-describeWithFlags('rand', ALL_ENVS, () => {
-  it('should return a random 1D float32 array', async () => {
-    const shape: [number] = [10];
-
-    // Enusre defaults to float32 w/o type:
-    let result = tf.rand(shape, () => util.randUniform(0, 2));
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 2);
-
-    result = tf.rand(shape, () => util.randUniform(0, 1.5));
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 1.5);
-  });
-
-  it('should return a random 1D int32 array', async () => {
-    const shape: [number] = [10];
-    const result = tf.rand(shape, () => util.randUniform(0, 2), 'int32');
-    expect(result.dtype).toBe('int32');
-    expectValuesInRange(await result.data(), 0, 2);
-  });
-
-  it('should return a random 1D bool array', async () => {
-    const shape: [number] = [10];
-    const result = tf.rand(shape, () => util.randUniform(0, 1), 'bool');
-    expect(result.dtype).toBe('bool');
-    expectValuesInRange(await result.data(), 0, 1);
-  });
-
-  it('should return a random 2D float32 array', async () => {
-    const shape = [3, 4];
-
-    // Enusre defaults to float32 w/o type:
-    let result = tf.rand(shape, () => util.randUniform(0, 2.5));
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 2.5);
-
-    result = tf.rand(shape, () => util.randUniform(0, 1.5), 'float32');
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 1.5);
-  });
-
-  it('should return a random 2D int32 array', async () => {
-    const shape = [3, 4];
-    const result = tf.rand(shape, () => util.randUniform(0, 2), 'int32');
-    expect(result.dtype).toBe('int32');
-    expectValuesInRange(await result.data(), 0, 2);
-  });
-
-  it('should return a random 2D bool array', async () => {
-    const shape = [3, 4];
-    const result = tf.rand(shape, () => util.randUniform(0, 1), 'bool');
-    expect(result.dtype).toBe('bool');
-    expectValuesInRange(await result.data(), 0, 1);
-  });
-
-  it('should return a random 3D float32 array', async () => {
-    const shape = [3, 4, 5];
-
-    // Enusre defaults to float32 w/o type:
-    let result = tf.rand(shape, () => util.randUniform(0, 2.5));
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 2.5);
-
-    result = tf.rand(shape, () => util.randUniform(0, 1.5), 'float32');
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 1.5);
-  });
-
-  it('should return a random 3D int32 array', async () => {
-    const shape = [3, 4, 5];
-    const result = tf.rand(shape, () => util.randUniform(0, 2), 'int32');
-    expect(result.dtype).toBe('int32');
-    expectValuesInRange(await result.data(), 0, 2);
-  });
-
-  it('should return a random 3D bool array', async () => {
-    const shape = [3, 4, 5];
-    const result = tf.rand(shape, () => util.randUniform(0, 1), 'bool');
-    expect(result.dtype).toBe('bool');
-    expectValuesInRange(await result.data(), 0, 1);
-  });
-
-  it('should return a random 4D float32 array', async () => {
-    const shape = [3, 4, 5, 6];
-
-    // Enusre defaults to float32 w/o type:
-    let result = tf.rand(shape, () => util.randUniform(0, 2.5));
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 2.5);
-
-    result = tf.rand(shape, () => util.randUniform(0, 1.5));
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 1.5);
-  });
-
-  it('should return a random 4D int32 array', async () => {
-    const shape = [3, 4, 5, 6];
-    const result = tf.rand(shape, () => util.randUniform(0, 2), 'int32');
-    expect(result.dtype).toBe('int32');
-    expectValuesInRange(await result.data(), 0, 2);
-  });
-
-  it('should return a random 4D bool array', async () => {
-    const shape = [3, 4, 5, 6];
-    const result = tf.rand(shape, () => util.randUniform(0, 1), 'bool');
-    expect(result.dtype).toBe('bool');
-    expectValuesInRange(await result.data(), 0, 1);
-  });
-});
-
 describeWithFlags('eye', ALL_ENVS, () => {
   it('1x1', async () => {
     const r = tf.eye(1);
@@ -944,143 +833,6 @@ describeWithFlags('eye', ALL_ENVS, () => {
     expect(r.dtype).toBe('bool');
     expect(r.shape).toEqual([3, 3]);
     expectArraysClose(await r.data(), [1, 0, 0, 0, 1, 0, 0, 0, 1]);
-  });
-});
-
-describeWithFlags('randomUniform', ALL_ENVS, () => {
-  it('should return a random 1D float32 array', async () => {
-    const shape: [number] = [10];
-
-    // Ensure defaults to float32 w/o type:
-    let result = tf.randomUniform(shape, 0, 2.5);
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 2.5);
-
-    result = tf.randomUniform(shape, 0, 1.5, 'float32');
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 1.5);
-  });
-
-  it('should return a random 1D int32 array', async () => {
-    const shape: [number] = [10];
-    const result = tf.randomUniform(shape, 0, 2, 'int32');
-    expect(result.dtype).toBe('int32');
-    expectValuesInRange(await result.data(), 0, 2);
-  });
-
-  it('should return a random 1D bool array', async () => {
-    const shape: [number] = [10];
-    const result = tf.randomUniform(shape, 0, 1, 'bool');
-    expect(result.dtype).toBe('bool');
-    expectValuesInRange(await result.data(), 0, 1);
-  });
-
-  it('should return a random 2D float32 array', async () => {
-    const shape: [number, number] = [3, 4];
-
-    // Ensure defaults to float32 w/o type:
-    let result = tf.randomUniform(shape, 0, 2.5);
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 2.5);
-
-    result = tf.randomUniform(shape, 0, 1.5, 'float32');
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 1.5);
-  });
-
-  it('should return a random 2D int32 array', async () => {
-    const shape: [number, number] = [3, 4];
-    const result = tf.randomUniform(shape, 0, 2, 'int32');
-    expect(result.dtype).toBe('int32');
-    expectValuesInRange(await result.data(), 0, 2);
-  });
-
-  it('should return a random 2D bool array', async () => {
-    const shape: [number, number] = [3, 4];
-    const result = tf.randomUniform(shape, 0, 1, 'bool');
-    expect(result.dtype).toBe('bool');
-    expectValuesInRange(await result.data(), 0, 1);
-  });
-
-  it('should return a random 3D float32 array', async () => {
-    const shape: [number, number, number] = [3, 4, 5];
-
-    // Ensure defaults to float32 w/o type:
-    let result = tf.randomUniform(shape, 0, 2.5);
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 2.5);
-
-    result = tf.randomUniform(shape, 0, 1.5, 'float32');
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 1.5);
-  });
-
-  it('should return a random 3D int32 array', async () => {
-    const shape: [number, number, number] = [3, 4, 5];
-    const result = tf.randomUniform(shape, 0, 2, 'int32');
-    expect(result.dtype).toBe('int32');
-    expectValuesInRange(await result.data(), 0, 2);
-  });
-
-  it('should return a random 3D bool array', async () => {
-    const shape: [number, number, number] = [3, 4, 5];
-    const result = tf.randomUniform(shape, 0, 1, 'bool');
-    expect(result.dtype).toBe('bool');
-    expectValuesInRange(await result.data(), 0, 1);
-  });
-
-  it('should return a random 4D float32 array', async () => {
-    const shape: [number, number, number, number] = [3, 4, 5, 6];
-
-    // Ensure defaults to float32 w/o type:
-    let result = tf.randomUniform(shape, 0, 2.5);
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 2.5);
-
-    result = tf.randomUniform(shape, 0, 1.5, 'float32');
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 1.5);
-  });
-
-  it('should return a random 4D int32 array', async () => {
-    const shape: [number, number, number, number] = [3, 4, 5, 6];
-    const result = tf.randomUniform(shape, 0, 2, 'int32');
-    expect(result.dtype).toBe('int32');
-    expectValuesInRange(await result.data(), 0, 2);
-  });
-
-  it('should return a random 4D bool array', async () => {
-    const shape: [number, number, number, number] = [3, 4, 5, 6];
-    const result = tf.randomUniform(shape, 0, 1, 'bool');
-    expect(result.dtype).toBe('bool');
-    expectValuesInRange(await result.data(), 0, 1);
-  });
-
-  it('should return a random 5D float32 array', async () => {
-    const shape: [number, number, number, number, number] = [2, 3, 4, 5, 6];
-
-    // Ensure defaults to float32 w/o type:
-    let result = tf.randomUniform(shape, 0, 2.5);
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 2.5);
-
-    result = tf.randomUniform(shape, 0, 1.5, 'float32');
-    expect(result.dtype).toBe('float32');
-    expectValuesInRange(await result.data(), 0, 1.5);
-  });
-
-  it('should return a random 5D int32 array', async () => {
-    const shape: [number, number, number, number, number] = [2, 3, 4, 5, 6];
-    const result = tf.randomUniform(shape, 0, 2, 'int32');
-    expect(result.dtype).toBe('int32');
-    expectValuesInRange(await result.data(), 0, 2);
-  });
-
-  it('should return a random 5D bool array', async () => {
-    const shape: [number, number, number, number, number] = [2, 3, 4, 5, 6];
-    const result = tf.randomUniform(shape, 0, 1, 'bool');
-    expect(result.dtype).toBe('bool');
-    expectValuesInRange(await result.data(), 0, 1);
   });
 });
