+++ tfjs-core/src/ops/ops.ts
@@ -17,8 +17,10 @@
 
 // Modularized ops.
 export {broadcastTo} from './broadcast_to';
+export {rand} from './rand';
 export {randomGamma} from './random_gamma';
 export {randomNormal} from './random_normal';
+export {randomUniform} from './random_uniform';
 export {square} from './square';
 export {squaredDifference} from './squared_difference';
 export {truncatedNormal} from './truncated_normal';
