+++ tfjs-core/src/tests.ts
@@ -78,6 +78,9 @@ import './ops/operation_test';
 import './ops/pad_test';
 import './ops/pool_test';
 import './ops/rand_test';
+import './ops/random_gamma_test';
+import './ops/random_normal_test';
+import './ops/random_uniform_test';
 import './ops/reduction_ops_test';
 import './ops/resize_bilinear_test';
 import './ops/resize_nearest_neighbor_test';
@@ -93,6 +96,7 @@ import './ops/spectral_ops_test';
 import './ops/strided_slice_test';
 import './ops/topk_test';
 import './ops/transpose_test';
+import './ops/truncated_normal_test';
 import './ops/unary_ops_test';
 import './optimizers/adadelta_optimizer_test';
 import './optimizers/adagrad_optimizer_test';
